[REST] Added a test data generator for unit testing the REST API
