+++ .vs/VSWorkspaceState.json
@@ -1,0 +1,9 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\flaskplayground",
+    "\\flaskplayground\\app"
+  ],
+  "SelectedNode": "\\flaskplayground\\upload.py",
+  "PreviewInSolutionExplorer": false
+}
