+++ flaskplayground/api/parsers.py
@@ -1,0 +1,255 @@
+from flask_restful import reqparse
+
+
+datetype = lambda x: datetime.strptime(x, '%Y-%m-%d')
+date_error_help = "Date fields should be entered as: YYYY-MM-DD"
+
+
+################ Parsers ####################
+
+
+# A base entity parser for other parsers to derive from
+base_parser = reqparse.RequestParser()
+base_parser.add_argument('english_name', required=True)
+base_parser.add_argument('chinese_name')
+base_parser.add_argument('pinyin_name')
+
+# Parser for input date related to a child object.
+child_parser = base_parser.copy()
+child_parser.add_argument('nickname')
+child_parser.add_argument('sex', required=True)
+child_parser.add_argument('birth_date', type=datetype, help=date_error_help)
+child_parser.add_argument('abandonment_date', type=datetype, help=date_error_help)
+child_parser.add_argument('program_entry_date', type=datetype, help=date_error_help)
+child_parser.add_argument('program_departure_date', type=datetype, help=date_error_help)
+child_parser.add_argument('program_departure_reason')
+child_parser.add_argument('child_history')
+child_parser.add_argument('medical_history')
+
+# When updating a child, no argument is required
+# so replace arguments fields from the original child_parser.
+child_update_parser = child_parser.copy()
+child_update_parser.replace_argument('english_name', required=False)
+child_update_parser.replace_argument('sex', required=False)
+
+# child_note
+child_note_parser = reqparse.RequestParser()
+child_note_parser.add_argument('date', type=datetype, help=date_error_help)
+child_note_parser.add_argument('note', required=True)
+child_note_parser.add_argument('flag', type=bool)
+child_note_parser.add_argument('child_id', required=True)
+
+child_note_update_parser = child_note_parser.copy()
+child_note_update_parser.replace_argument('note', required=False)
+child_note_update_parser.replace_argument('child_id', required=False)
+
+# partner
+partner_parser = base_parser.copy()
+partner_parser.add_argument('email')
+partner_parser.add_argument('phone')
+
+partner_update_parser = partner_parser.copy()
+partner_update_parser.replace_argument('english_name', required=False)
+
+# caregiver
+caregiver_parser = base_parser.copy()
+
+caregiver_update_parser = caregiver_parser.copy()
+caregiver_update_parser.replace_argument('english_name', required=False)
+
+# specialist
+
+specialist_parser = base_parser.copy()
+specialist_parser.add_argument('specialist_type_id', type=int, required=True)
+specialist_update_parser = specialist_parser.copy()
+for arg in specialist_update_parser.args:
+    specialist_update_parser.replace_argument(arg, required=False)
+
+# specialist_type
+
+specialist_type_parser = base_parser.copy()
+specialist_type_update_parser = specialist_type_parser.copy()
+for arg in specialist_type_update_parser.args:
+    specialist_type_update_parser.replace_argument(arg, required=False)
+
+# milestone_type_category
+
+milestone_type_category_parser = base_parser.copy()
+milestone_type_category_update_parser = milestone_type_category_parser.copy()
+for arg in milestone_type_category_update_parser.args:
+    milestone_type_category_update_parser.replace_argument(arg, required=False)
+
+# milestone_type
+
+milestone_type_parser = base_parser.copy()
+milestone_type_update_parser = milestone_type_parser.copy()
+for arg in milestone_type_update_parser.args:
+    milestone_type_update_parser.replace_argument(arg, required=False)
+
+# doctor_type
+
+doctor_type_parser = base_parser.copy()
+doctor_type_update_parser = doctor_type_parser.copy()
+for arg in doctor_type_update_parser.args:
+    doctor_type_update_parser.replace_argument(arg, required=False)
+
+# doctor
+
+doctor_parser = reqparse.RequestParser()
+doctor_parser.add_argument('doctor_english_name', required=True)
+doctor_parser.add_argument('doctor_chinese_name')
+doctor_parser.add_argument('doctor_pinyin_name')
+doctor_parser.add_argument('facility_english_name')
+doctor_parser.add_argument('facility_chinese_name')
+doctor_parser.add_argument('facility_pinyin_name')
+doctor_parser.add_argument('doctor_type_id', type=int, required=True)
+doctor_update_parser = doctor_parser.copy()
+for arg in doctor_update_parser.args:
+    doctor_update_parser.replace_argument(arg, required=False)
+
+# measurement_type
+
+measurement_type_parser = base_parser.copy()
+measurement_type_parser.add_argument('units', required=True)
+measurement_type_update_parser = measurement_type_parser.copy()
+for arg in measurement_type_update_parser.args:
+    measurement_type_update_parser.replace_argument(arg, required=False)
+
+# camp
+
+camp_parser = base_parser.copy()
+camp_update_parser = camp_parser.copy()
+for arg in camp_update_parser.args:
+    camp_update_parser.replace_argument(arg, required=False)
+
+# medical_condition
+
+medical_condition_parser = base_parser.copy()
+medical_condition_update_parser = medical_condition_parser.copy()
+for arg in medical_condition_update_parser.args:
+    medical_condition_update_parser.replace_argument(arg, required=False)
+
+# medication
+
+medication_parser = base_parser.copy()
+medication_parser.add_argument('milligram_dose', type=float, required=True)
+medication_update_parser = medication_parser.copy()
+for arg in medication_update_parser.args:
+    medication_update_parser.replace_argument(arg, required=False)
+
+# child_partner
+child_partner_parser = reqparse.RequestParser()
+child_partner_parser.add_argument('start_date', type=datetype, help=date_error_help)
+child_partner_parser.add_argument('end_date', type=datetype, help=date_error_help)
+child_partner_parser.add_argument('note')
+child_partner_parser.add_argument('flag', type=bool)
+child_partner_parser.add_argument('child_id', type=int, required=True)
+child_partner_parser.add_argument('partner_id', type=int, required=True)
+
+child_partner_update_parser = child_partner_parser.copy()
+child_partner_update_parser.replace_argument('child_id', type=int, required=False)
+child_partner_update_parser.replace_argument('partner_id', type=int, required=False)
+
+# child_camp
+child_camp_parser = reqparse.RequestParser()
+child_camp_parser.add_argument('date', required=True, type=datetype, help=date_error_help)
+child_camp_parser.add_argument('note')
+child_camp_parser.add_argument('child_id', type=int, required=True)
+child_camp_parser.add_argument('camp_id', type=int, required=True)
+
+child_camp_update_parser = child_camp_parser.copy()
+child_camp_update_parser.replace_argument('date', required=False, type=datetype, help=date_error_help)
+child_camp_update_parser.replace_argument('child_id', type=int, required=False)
+child_camp_update_parser.replace_argument('camp_id', type=int, required=False)
+
+# child_assessment
+
+child_assessment_parser = reqparse.RequestParser()
+child_assessment_parser.add_argument('date', required=True, type=datetype, help=date_error_help)
+child_assessment_parser.add_argument('note')
+child_assessment_parser.add_argument('child_id', type=int, required=True)
+child_assessment_parser.add_argument('flag', type=bool)
+child_assessment_parser.add_argument('specialist_id', type=int, required=True)
+
+child_assessment_update_parser = child_assessment_parser.copy()
+child_assessment_update_parser.replace_argument('date', required=False, type=datetype, help=date_error_help)
+child_assessment_update_parser.replace_argument('child_id', type=int, required=False)
+child_assessment_update_parser.replace_argument('specialist_id', type=int, required=False)
+
+# child_caregiver
+
+child_caregiver_parser = reqparse.RequestParser()
+child_caregiver_parser.add_argument('start_date', type=datetype, help=date_error_help)
+child_caregiver_parser.add_argument('end_date', type=datetype, help=date_error_help)
+child_caregiver_parser.add_argument('note')
+child_caregiver_parser.add_argument('child_id', type=int, required=True)
+child_caregiver_parser.add_argument('caregiver_id', type=int, required=True)
+
+child_caregiver_update_parser = child_caregiver_parser.copy()
+child_caregiver_update_parser.replace_argument('child_id', type=int, required=False)
+child_caregiver_update_parser.replace_argument('caregiver_id', type=int, required=False)
+
+# child_measurement
+
+child_measurement_parser = reqparse.RequestParser()
+child_measurement_parser.add_argument('date', required=True, type=datetype, help=date_error_help)
+child_measurement_parser.add_argument('value', required=True, type=float)
+child_measurement_parser.add_argument('child_id', type=int, required=True)
+child_measurement_parser.add_argument('measurement_type_id', type=int, required=True)
+
+child_measurement_update_parser = child_measurement_parser.copy()
+child_measurement_update_parser.replace_argument('date', required=False, type=datetype, help=date_error_help)
+child_measurement_update_parser.replace_argument('child_id', type=int, required=False)
+child_measurement_update_parser.replace_argument('measurement_type_id', type=int, required=False)
+child_measurement_update_parser.replace_argument('value', required=False, type=float)
+
+# child_milestone
+
+child_milestone_parser = reqparse.RequestParser()
+child_milestone_parser.add_argument('date', required=True, type=datetype, help=date_error_help)
+child_milestone_parser.add_argument('child_id', type=int, required=True)
+child_milestone_parser.add_argument('milestone_type_id', type=int, required=True)
+
+child_milestone_update_parser = child_milestone_parser.copy()
+child_milestone_update_parser.replace_argument('date', required=False, type=datetype, help=date_error_help)
+child_milestone_update_parser.replace_argument('child_id', type=int, required=False)
+child_milestone_update_parser.replace_argument('milestone_type_id', type=int, required=False)
+
+# child_doctor_visit
+
+child_doctor_visit_parser = reqparse.RequestParser()
+child_doctor_visit_parser.add_argument('date', required=True, type=datetype, help=date_error_help)
+child_doctor_visit_parser.add_argument('child_id', type=int, required=True)
+child_doctor_visit_parser.add_argument('doctor_id', type=int, required=True)
+child_doctor_visit_parser.add_argument('note')
+
+child_doctor_visit_update_parser = child_doctor_visit_parser.copy()
+child_doctor_visit_update_parser.replace_argument('date', required=False, type=datetype, help=date_error_help)
+child_doctor_visit_update_parser.replace_argument('child_id', type=int, required=False)
+child_doctor_visit_update_parser.replace_argument('doctor_id', type=int, required=False)
+
+# child_medical_condition
+
+child_medical_condition_parser = reqparse.RequestParser()
+child_medical_condition_parser.add_argument('child_id', type=int, required=True)
+child_medical_condition_parser.add_argument('medical_condition_id', type=int, required=True)
+
+child_medical_condition_update_parser = child_medical_condition_parser.copy()
+child_medical_condition_update_parser.replace_argument('child_id', type=int, required=False)
+child_medical_condition_update_parser.replace_argument('medical_condition_id', type=int, required=False)
+
+# child_medication
+
+child_medication_parser = reqparse.RequestParser()
+child_medication_parser.add_argument('start_date', required=True, type=datetype, help=date_error_help)
+child_medication_parser.add_argument('end_date', type=datetype, help=date_error_help)
+child_medication_parser.add_argument('dosage1', type=float)
+child_medication_parser.add_argument('dosage2', type=float)
+child_medication_parser.add_argument('dosage3', type=float)
+child_medication_parser.add_argument('child_id', type=int, required=True)
+child_medication_parser.add_argument('medication_id', type=int, required=True)
+
+child_medication_update_parser = child_medication_parser.copy()
+child_medication_update_parser.replace_argument('start_date', required=False, type=datetype, help=date_error_help)
+child_medication_update_parser.replace_argument('child_id', type=int, required=False)
+child_medication_update_parser.replace_argument('medication_id', type=int, required=False)
