+++ flaskplayground/app/__init__.py
@@ -9,13 +9,13 @@ api = Api(app)
 
 import models
 
-from resources import QueryResource
-from resources import EntityResource
-from resources import EntityListResource
-from resources import EntityFilterResource
-from resources import HeartbeatResource
-from resources import RollbackResource
-from upload import Upload
+from api.resources import QueryResource
+from api.resources import EntityResource
+from api.resources import EntityListResource
+from api.resources import EntityFilterResource
+from api.resources import HeartbeatResource
+from api.resources import RollbackResource
+from api.upload import Upload
 
 api.add_resource(QueryResource, "/query", endpoint="query")
 api.add_resource(EntityResource, "/entity/<string:entity_name>", endpoint="entity")
