--- flaskplayground/app/forms.py
@@ -1,7 +1,0 @@
-#from flask_wtf import FlaskForm
-#from wtforms import StringField, BooleanField
-#from wtforms.validators import DataRequired
-#
-#class LoginForm(FlaskForm):
-#    openid = StringField('openid', validators=[DataRequired()])
-#    remember_me = BooleanField('remember_me', default=False)
